refactor(ghostmap): split overlap-schedule predicates into GhostMapPresence.OverlapSchedule.cs
